Use a suitable include path with matching includes to find Detour headers.
This fixes compile errors when building contrib/mmaps with CMake:
src/game/MoveMapSharedDefines.h:23:62: fatal error:
  ../recastnavigation/Detour/Include/DetourNavMesh.h: No such file or directory

Note that the core is able to use this header without any issues because it sets
up dep/include as in include path (see src/game/CMakeLists.txt). This means the
compiler will be able to find the relative include anyway, but it's relying on
an irrelevant include path - there's nothing related to Recast or Detour in
dep/include - that contrib/mmaps does not have.