Fix contrib/mmaps and contrib/vmap_assembler: build the ACE libraries if needed.
Building contrib/mmaps and contrib/vmap_assembler requires compiling (vmap)
source files, which depend on ACE headers. On Unix-like systems, there is an
option to build the ACE library included in the source tree instead of using
an external ACE library. In this case, some ACE headers needed by the vmap
sources (such as ace/config.h) are not available without building ACE first,
and merely setting up an include path is not sufficient. Therefore, we need
to build ACE first in this case.